Keep the wizard at three steps: Customize is one collapsed row, not a step
The previous commit made the four settings a mandatory fourth step, which
is the shape #41 explicitly rules out: "defaults must stay one click
away, since most people should not have to make these choices."

They now live behind a collapsed "Customize this plan" row at the bottom
of the schedule step, above the summary. Closed — which is how everyone
who does not want them ever sees it — the wizard is the three-step flow
it was, with one extra line of text and no extra decisions or clicks.
Opened, it shows the same four sections Edit Plan does, from the same
shared views.

The row is hand-rolled rather than a DisclosureGroup, matching the
destination step's Advanced options: the whole two-line title+subtitle is
the click target, not just the chevron (issue #50). The step scrolls only
once it is opened, so the window goes back to 560x560.
